Fix Lexer not parsing first function.
If a function declaration is the first line of of the file
the lexer can not put a `DeDent` token do to how it detects
function declaration with `\n` char. To fix this I just
insert a `\n` at the start of all files being lexed.

The parser uses the `DeDent` as a fn or def in front of the function
declaration.